added middleware for some routes

+++ app/api/comment/[id]/route.ts
@@ -1,26 +1,35 @@
 import prisma from "@/lib/prisma";
 import { NextRequest, NextResponse } from "next/server";
+import { isAuthenticated } from "../../isAuthenticated";
 
 export async function POST(
   req: NextRequest,
   { params }: { params: { id: string } }
 ) {
   try {
-    const Body = await req.json();
-    const { body, userId } = Body;
+    if (isAuthenticated(req)) {
+      const Body = await req.json();
+      const { body, userId } = Body;
 
-    const comment = await prisma.comment.create({
-      data: {
-        body,
-        post: {
-          connect: { id: params.id },
+      const comment = await prisma.comment.create({
+        data: {
+          body,
+          post: {
+            connect: { id: params.id },
+          },
+          user: {
+            connect: { id: userId },
+          },
         },
-        user: {
-          connect: { id: userId },
-        },
-      },
-    });
-    return NextResponse.json(comment, { status: 201 });
+      });
+
+      return NextResponse.json(comment, { status: 201 });
+    } else {
+      return NextResponse.json(
+        { message: "Authentication failed" },
+        { status: 401 }
+      );
+    }
   } catch (error) {
     console.error(error);
     return NextResponse.json(
@@ -35,13 +44,19 @@ export async function DELETE(
   { params }: { params: { id: string } }
 ) {
   try {
-    console.log("params", params);
-    const deleteComment = await prisma.comment.delete({
-      where: {
-        id: params.id,
-      },
-    });
-    return NextResponse.json({ message: "Comment deleted" }, { status: 200 });
+    if (isAuthenticated(req)) {
+      const deleteComment = await prisma.comment.delete({
+        where: {
+          id: params.id,
+        },
+      });
+      return NextResponse.json({ message: "Comment deleted" }, { status: 200 });
+    } else {
+      return NextResponse.json(
+        { message: "Authentication failed" },
+        { status: 401 }
+      );
+    }
   } catch (error) {
     console.error(error);
     return NextResponse.json(
